+++ src/core/backup/sram.cpp
@@ -4,21 +4,20 @@
 #include "sram.hpp"
 #include "gba.hpp"
 #include "mem.hpp"
-#include <algorithm>
-#include <ranges>
+#include <cstring>
 
 namespace gba::backup::sram {
 
 auto Sram::init([[maybe_unused]] Gba& gba) -> void
 {
-    std::ranges::fill(this->data, 0xFF);
+    std::memset(this->data, 0xFF, sizeof(this->data));
 }
 
 auto Sram::load_data(std::span<const u8> new_data) -> bool
 {
     if (new_data.size() <= std::size(this->data))
     {
-        std::ranges::copy(new_data, this->data);
+        std::memcpy(this->data, new_data.data(), new_data.size());
         return true;
     }
     else